remodify useAddEmployee logic
